add babel, restructure files, update readme
