+++ .babelrc
@@ -1,0 +1,14 @@
+{
+  "presets": [
+    "@babel/preset-react",
+    [
+      "@babel/preset-env",
+      {
+        "targets": {
+          "edge": "14"
+        },
+        "modules": "commonjs"
+      }
+    ]
+  ]
+}
