+++ src/index.js
@@ -1,0 +1,9 @@
+import SVGIcon from './svg-icon';
+import FionaPlaceholder from './FionaPlaceholder';
+
+export default {
+  title: 'Fiona Placeholder',
+  name: 'fionaPlaceholder',
+  icon: SVGIcon,
+  component: FionaPlaceholder,
+};
